Function User and Purchase History

+++ src/main/java/com/group1/FMobile/domain/OrdersDetail.java
@@ -34,6 +34,6 @@ public class OrdersDetail {
     Integer quantity;
 
     @Column(name = "price", nullable = false)
-    double price;
+    Long price;
 
 }
